Propagate HTTP errors upward to ease debugging.
In a few instances this semester, the JSON endpoint went down, returning 5xx errors for a few hours. The web team think it might be due to a security appliance in front of the webserver. I've asked them to whitelist the Mary's user-agent, but in the meantime raise any HTTP errors we get so issues are easier to trace for end users.